Refactor classpath and analysis lookups.
Introduce a `Lookup` trait that contains all necessary methods that are
needed by zinc to perform classpath and Analysis lookups for external
dependencies.

Adapt existing logic to make use of the newly introduced interface.
We have several different variants of `lookupAnalysis` because different
parts of zinc start either with just a binary class name or a class file
or both. Eventually we should perform all lookups based just on a binary
class name.