Fix naming of inherited classes in ExtractAPI
For an inherited class, ExtractAPI would form a (source) class name by
calling `Symbol.className` on the inherited class. However, that created
a name of a class as seen at declaration site and not at inheritance site.
Let's consider an example:

class A { class AA }
class B extends A

Before this change, ExtractAPI would create an API representation of `AA`
twice: once seen from A, and then the second time seen from B as an
inherited member. However, in both cases it would use `A.AA` as a name.
This commit fixes naming so an inherited representation of `AA` has
a name `B.AA`.